Feat: single mcp server for all integrations
