+++ apps/webapp/package.json
@@ -4,11 +4,11 @@
   "sideEffects": false,
   "type": "module",
   "scripts": {
-    "build": "remix vite:build",
-    "dev": "node ./server.mjs",
+    "build": "remix vite:build && tsc server.ts --outDir ./ --module ESNext --moduleResolution bundler --target ES2022 --allowSyntheticDefaultImports --skipLibCheck",
+    "dev": "tsx watch server.ts",
     "lint": "eslint --fix --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
     "lint:fix": "eslint 'app/**/*.{ts,tsx,js,jsx}' --rule 'turbo/no-undeclared-env-vars:error' -f table",
-    "start": "remix-serve ./build/server/index.js",
+    "start": "node server.js",
     "typecheck": "tsc",
     "trigger:dev": "pnpm dlx trigger.dev@4.0.0-v4-beta.22 dev",
     "trigger:deploy": "pnpm dlx trigger.dev@4.0.0-v4-beta.22 deploy"
@@ -175,7 +175,8 @@
     "tailwindcss": "4.1.7",
     "typescript": "5.8.3",
     "vite": "^6.0.0",
-    "vite-tsconfig-paths": "^4.2.1"
+    "vite-tsconfig-paths": "^4.2.1",
+    "tsx": "4.20.4"
   },
   "engines": {
     "node": ">=20.0.0"
