+++ hosting/docker/.env
@@ -1,4 +1,4 @@
-VERSION=0.1.17
+VERSION=0.1.18
 
 # Nest run in docker, change host to database container name
 DB_HOST=postgres
