+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "core",
   "private": true,
-  "version": "0.1.17",
+  "version": "0.1.18",
   "workspaces": [
     "apps/*",
     "packages/*"
